added search and refactor
