+++ src/html/header.html
@@ -2,14 +2,11 @@
    <div class="container header">
       <div class="site-logo">
          <a href="#" class="site-logo__link">
-            {{!-- <svg width="24" height="24"> --}}
                <svg class="icon-film" width="24" height="24" fill="none" xmlns="http://www.w3.org/2000/svg">
                   <path
                      d="M19.82 2H4.18A2.18 2.18 0 002 4.18v15.64A2.18 2.18 0 004.18 22h15.64A2.18 2.18 0 0022 19.82V4.18A2.18 2.18 0 0019.82 2zM7 2v20M17 2v20M2 12h20M2 7h5M2 17h5M17 17h5M17 7h5"
                      stroke="#818181" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" />
                </svg>
-               {{!-- <use class="icon-film" href="./images/sprite.svg#icon-film-1"></use> --}}
-               {{!-- </svg> --}}
             <span class="site-logo__txt">Filmoteka</span>
          </a>
          <ul class="header-nav__list">
@@ -26,9 +23,6 @@
             <input class="menu-search__films search-field" type="text" name="text" autocomplete="off"
                placeholder="Поиск фильмов" />
             <button type="submit" class="search-btn">
-               <svg width="12" height="12">
-                  <use href="./images/sprite.svg#icon-search-2-1"></use>
-               </svg>
             </button>
          </form>
       </div>
