+++ src/index.js
@@ -1,6 +1,4 @@
 import './sass/main.scss';
 import './js/modal-window-film';
-
-import './js/render-popular';
-
-import './js/pages';
+import './js/header';
+import './js/render-main';
