+++ src/js/header.js
@@ -1,0 +1,7 @@
+import header from '../html/header.html';
+
+const wrapperRef = document.querySelector('.wrapper');
+
+wrapperRef.insertAdjacentHTML('afterbegin', header);
+
+// console.log(markup);
